tdf#167450 adapt sheet index in anchor to new pos
When a sheet is copied to another document, the tab index might change.
The tab index was not adapted in the anchor info in the shape, that are
maStart and maEnd in m_pPlusData in SdrObject. That resulted in a wrong
shape position after save and reload.

Change-Id: If69dcc9067e1346e1e115adf3c5ca68a9423e250
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/187802
Tested-by: Jenkins
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>
(cherry picked from commit 704dd0d279af6c1ace1bde61fec93e7c3a8d1035)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/187924
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>