tdf#80662 Fix display of JDK version number / path
Previously, JDK version number and path were displayed as RTL in right
to left mode of LibreOffice, in "Tools > Options > Advanced". The RTL UI
can be activated by setting:

SAL_RTL_ENABLED=1

The fix uses an LRM and PDF Unicode characters around the text:
https://en.wikipedia.org/wiki/Bidirectional_text

This patch also fixes the display of JDK path, which had the same issue
that was visible in the bug doc.

Change-Id: I80b25821defd43355e5b271d0348dd70ef520440
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/187813
Tested-by: Jenkins
Reviewed-by: Hossein   <hossein@libreoffice.org>