tdf#167007 Add animated graphic to cache when prepared
For some reason, after an animation has been swapped out by
MemoryManager::loopAndReduceMemory(), the animation repeatedly
creates a new ImpGraphic instance, swaps it in, but it never
gets registered in the cache. Since it is not in the cache, new
ImpGraphic instances get deleted almost immediately after they
are created.

So prevent immediate deletion by ensuring that animated
ImpGraphic instances are registered when they are prepared.

Change-Id: Ib721e0bc31d3eddc4ed4131d9d8dc8919eed7b44
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/187836
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Tested-by: Jenkins