tdf#167222 sw floattable: allow split of fly and its keep-together anchor text
Regression from commit 3eb6d764b3023500f2299d36bf1860bc8e67db9f (DOCX
import: floating table with negative top margin has to be a fly frame,
2022-01-21), open the bugdoc, notice that the floating table should be
at the end of page 1 but is at the start of page 2.

This happens because in general keep-together means no split should
happen for all anchored objects and lines of a text frame, but Word
allows floating tables to be split from the next regular paragraph in
the document.

Fix the problem by building on top of the work done in commit
2d0a4ef1d83b8de6cb133970c2c35ae706fb058e (sw floattable: fix negative
vertical offset handling on page boundary, 2023-06-20), which makes this
split to "floating table" and "anchor text" pieces possible, but that
didn't help here, since the anchor text node happened to be a heading,
with both a split property set to false and a keep-with-next property
set to true. Now we ignore both properties in case the paragraph is a
floating table anchor, which means that in SwTextFrame::FormatAdjust(),
rFrameBreak.IsBreakNow() can return true, resulting in a call to
SplitFrame(), as needed.

The behavior for other types of anchored objects is unchanged, those
don't go to a separate page in Word, either.

Change-Id: I48835f355fa5aeb09b1a92bd8cee7e06cd148240
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/187858
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins