ScriptForge (Form) Fix missing attributes
When a FormControl instance is created
triggered by an event with
   fc = CreateScriptService("FormEvent", event)
a number of attributes are not
(re)-initialized in the parent Form instance.

They concern references to the form document
container and to the database connetction.

The absences prevent from scripting easily
(workarounds exist) operations on the form
document (close, print, ...) or towards the
database (read, sql, ...). In particular,
this limits the reusability of scripts
across >1 form documents.

This is a bug. It is fixed with the actual change.

Documentation may be left unchanged.
Basic and Python scripts are impacted alike.

Change-Id: I285148917174d5a046c11b15ea2a2581a23d4277
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/187854
Tested-by: Jenkins
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>