tdf#166319 sw interdependent redlines: handle deleting a self ins-then-fmt
The bugdoc has <ins>A<format>B</format>C</ins>, go to the end, delete C,
result is <ins>A<format>B</format></ins> given that this was a
self-insert, but the next delete results in <ins>A</ins><del>B</del>.

The problem is that we have special code to deal with delete on
self-insert (it's fine to just remove such content from the model), but
this code won't handle insert-then-format as insert, while Word does.

Fix the problem by extending
DocumentRedlineManager::PreAppendDeleteRedline() to look "through" just
format redlines to see the underlying insert.

Note that the same is not needed for delete-then-format, since insert is
always inserted next to an existing content or redlined content, so
there is no overlap.

Change-Id: I8b47839ffaa470714a5dd37722aa26ac0002c0bc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/187907
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins
(cherry picked from commit f66d63da05dbe2f254ffaf428257684a38523f66)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/187921
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>