reportdesign: Create XML exporters without UNO ctors
So far, the XML exporters used in
OReportDefinition::storeToStorage and
OReportDefinition::WriteThroughComponent
were created using UNO service constructors.

Move away from that approach/indirection:
Add static helper methods to ORptExport
that can be used to create ORptExport
instances with the relevant parameters
(that were previously passed to the ctor
by the UNO service constructors) and call
these directly to receive the relevant instances.

Drop the now unused UNO constructors.

Pass the `rtl::Reference<rptxml::ORptExport>&`
to OReportDefinition::WriteThroughComponent,
which also avoids the need to query for the relevant
interfaces.

No change in user-observable behavior intended
or seen in a quick test of generating a report
in Base.

Change-Id: Ia84319009e71ef432077a60dd7e44003126334d6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/187914
Tested-by: Jenkins
Reviewed-by: Adam Seskunas <adamseskunas@gmail.com>
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>