tdf#166734 CairoSDPR: use sub-pixel correct transformation
When parts of the output need to be pre-rendered (e.g.
transparencies) the temp target is of course pixel-aligned,
bit the mapping for the rendering was not yet. Added that
now. This could lead to sub-pixel displacements which are
usually not recognizable, but also like here make the
PixelSnap fail.

Change-Id: I877987947bdcc7521580de9287894ca70b17763c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/187928
Reviewed-by: Armin Le Grand <Armin.Le.Grand@me.com>
Tested-by: Jenkins
(cherry picked from commit 15a5d633fec3c3c62870ab4168a1886796350031)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/187958
Reviewed-by: Ilmari Lauhakangas <ilmari.lauhakangas@libreoffice.org>