reportdesign: Use ref to concrete OReportDefinition
The XReportDefinition used in ORptExport
(set by calling ORptExport::setSourceDocument
from OReportDefinition::WriteThroughComponent)
is always an OReportDefinition.

Use an rtl::Reference to that class instead
of a reference to the abstract XReportDefinition.

Background is to allow extending OReportDefinition
as needed without having to adjust the XReportDefinition
interface in the context of the
"Implement ReportBuilder in C++" GSoC project [1]
in the future.

[1] https://lists.freedesktop.org/archives/libreoffice/2025-May/093278.html

Change-Id: I5780fe7d7a809183d4dfb0018b37d6f2ed2ee36b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/187948
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>