Introduce o3tl::getEnvironment
Getting environment variable values into OUString is a common task.
In many places, we either use getenv and convert to OUString; or
use osl_getEnvironment (usually ignoring its return value), which
is a bit awkward.

The introduced o3tl::getEnvironment is a wrapper around the latter.
One of upsides of using it is minimizing direct access to pData.

o3tl::setEnvironment is also added for symmetry.

Change-Id: I420093de4873cc08c8b64d218ab93218d8ad398f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/187973
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins