weld: Use more specific CheckButton param for TriStateEnabled
The TriStateEnabled struct implements logic to loop
through the three different TriState states for a
weld::Toggleable.

In practice, it is only used for weld::CheckButton
objects, and (while not instantly clear from the API)
TRISTATE_INDET isn't actually supported
by all weld::Toggleable subclasses.

Switch the param to TriStateEnabled::ButtonToggled
from weld::Toggleable& to weld::CheckButton& and
rename to CheckButtonToggled to make clearer this
is only for CheckButton.

As a consequence, adjust callers that were passing
just a weld::Toggleable& to pass a CheckButton&
(to the same underlying object) to adhere to the
new method signature.

This is one step towards making the inconsistent state
specific to weld::CheckButton, i.e. (re)move it from
weld::Toggleable and other subclasses.

Change-Id: I843c1c7075f91c0b81605ecb5d7e54621b9074c1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/187992
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>