Related: tdf#167222 sw floattable: fix split of fly and its 'keep' anchor text
Open the bugdoc, the floating table is at the bottom of page 2 in Word,
but it's at the top of page 3 in Writer.

This is similar to commit b8009818d619bff16d19a0f589a733f78a46ec0b
(tdf#167222 sw floattable: allow split of fly and its keep-together
anchor text, 2025-07-14), but here an SwTextFrameBreak instance is
created, IsBreakNow() is invoked multiple times: so by the time the
first IsBreakNow() runs, m_bKeep is false, but later it gets turned on.

Fix the problem by adapting SwTextFrameBreak::IsBreakNow() to be
consistent with its ctor, which already turns off m_bKeep for split fly
anchors. This makes sense, because otherwise it can happen that first
IsBreakNow() is invoked with a text frame where the format is in
progress and we're not yet ready to split, but later we could split, but
we don't even try due to the changed m_bKeep flag.

Change-Id: I5d122e60c7693614e6e4c1b90e83f622be18fb72
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/187983
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>