svx: Extract TriStateToggleButton base from LightButton
Svx3DWin is the only place where weld::ToggleButton
is currently used with TRISTATE_INDET to indicate
a special "inconsistent"/invalid state.
That state is not supported by QtInstanceToggleButton
and for GTK 4, GtkToggleButton also no longer directly
supports that as it did in GTK 3, see GTK commit [1]

    commit 2aea8dfee97b514e1fca7f577fefdac4e449e906
    Author: Timm Bäder <mail@baedert.org>
    Date:   Sat Jan 28 08:11:14 2017 +0100

        togglebutton: Move :inconsistent to GtkCheckButton

In order to make the inconsistent state specific to
weld::CheckButton and dropping weld::Toggleable::set_state
etc. in an upcoming commit, the relevant logic to
have that special indeterminate state still available
for Svx3DWin, introduce TriStateToggleButton used only
here to still support the use in this one use case.
(There might be a better approach altogether here, too,
but I'm not familiar enough with the underlying logic,
so try to keep the general logic unchanged for now.)

This will be used for more ToggleButtons in Svx3DWin
that also make use of TRISTATE_INDET.

[1] https://gitlab.gnome.org/GNOME/gtk/-/commit/2aea8dfee97b514e1fca7f577fefdac4e449e906

Change-Id: I4b39f941fdda886390079a00aed4c2ce29be5a43
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/187999
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins