tdf#130857 svx Svx3DWin: Stop relying on weld::ToggleButton indeterminate state
Stop relying on weld::ToggleButton supporting
TRISTATE_INDET in the TriStateToggleButton helper
class, but instead introduce TriStateToggleButton::m_bIndeterminate
to explicitly get/set that state in the helper class
to indicate a "no real value set" state,
in preparation of removing support for TRISTATE_INDET
from weld::ToggleButton in an upcoming commit.

See also previous commit

    Change-Id: I4b39f941fdda886390079a00aed4c2ce29be5a43
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Jul 17 10:31:06 2025 +0200

        svx: Extract TriStateToggleButton base from LightButton

for more background.

No change in behavior intended/expected for that
dialog, but (independent of this change) I didn't see
this "indeterminate" code path being taken in practice in
a quick test of that dialog started as follows:

* start Impress
* enable "View" -> "Toolbars" -> "3D-Objects"
* click on any object in the toolbar (e.g. "Cube")
  and insert it into the slide
* right-click on the object and select the
  "3D Dffects" context menu entry

Change-Id: Ib781e9108d526f90282a203b53065111a9d12236
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188001
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>