reportdesign: Use concrete OReportDefinition in OReportEngineJFree
The abstract XReportDefinition interface param allows calling
OReportEngineJFree::setReportDefinition
via the XReportEngine interface from dbaccess
(s. ODocumentDefinition::onCommandOpenSomething)
which is lower in the dependency chain than reportdesign, i.e.
using the UNO abstraction helps to avoid a dependency cycle.

However, OReportDefinition is the only implementation of that
XReportDefinition interface actually used and can be used
directly inside OReportEngineJFree.

Therefore, cast to the specific class and use an rtl::Reference
instead of a reference to the abstract XReportDefinition.

Background is to allow extending OReportDefinition
as needed without having to adjust the XReportDefinition
interface in the context of the
"Implement ReportBuilder in C++" GSoC project [1]
in the future.

[1] https://lists.freedesktop.org/archives/libreoffice/2025-May/093278.html

Change-Id: Ia442dd4f64bcf2bf3ba04707ef8efcade932c2ff
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188014
Tested-by: Jenkins
Reviewed-by: Adam Seskunas <adamseskunas@gmail.com>
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>