tdf#130857 qt weld: Support "Text Attributes" dialog
This means that native Qt widgets are used for that dialog
now when using the qt5 or qt6 VCL plugin and starting LO with
environment variable SAL_VCL_QT_USE_WELDED_WIDGETS=1 set.

To trigger the dialog:

* create new Impress presentation
* select the title frame
* right-click to open context menu
* select the "Text Attributes..." context entry

Currently, there is a crash when closing the dialog
after interacting with the UI elements, but that
is a preexisting issue that will be addressed in
a separate upcoming commit.

Change-Id: I2e9d9a4c24251c0a0119cb6a9133ad02b70ebee0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188031
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>