tdf#167591: Only apply date correction, when sDefault wasn't date string
The correction is needed to adjust a date in the data source's epoch
to the epoch used by the control (and its formatter). These epochs
may differ.

The problem was, that when sDefault has a date string like "20/07/25"
or "2025-07-20", the conversion from that string to double using the
control's formatter already created the value in the required epoch;
and the correction broke the value.

Fix that by checking if the string was really a date string, and so
the correction should be skipped.

Change-Id: Ia1ba2b1b5dc88d66d524ff03fecc028f7ebf3336
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188091
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>