Related: tdf#167591 Make STANDARD_DB_DATE more standard
For some reason, STANDARD_DB_DATE was 1900-01-01, which differs from
the most used epoch of 1899-12-30 (see e.g. the default initializer
of ImpSvNumberformatScan). This was triggering tdf#167591, where the
correction added two days of the difference to the date entered into
the default value box.

I couldn't find how/if this value could get exposed to API; so let's
try to standardise it a bit.

Change-Id: I9207cbe3a5a91ca63f3f97792da0c5ad2931ffa8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188092
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins