tdf#130857 Use explicit column for image in formula dlg "Struct" page
Use a separate column (at index 0) for the image shown
for entries in the treeview in the "Structure" page
of the formula dialog instead of using the index of -1
which is documented to set the "expander image" when
passed to weld::TreeView::set_image.

There is however a separate icon used to actually expand/collapse
the entry, and both, the GtkInstanceTreeView and the
SalInstanceTreeView implementation internally use
a separate extra column index for that image internally
instead.

Stop relying on that special logic for this tree view
and use a "normal" column instead, and using only a single
GtkCellRenderer per column.

This will also simplify supporting this for the Qt
implementation in QtInstanceTreeView, which doesn't
support the special "expander column" at index -1 at
this point in time.

No change in behavior expected or seen in a quick
test of that dialog in Calc:

* "Insert" -> "Function"
* switch to the "Structure" tab

Change-Id: I80e191305930f5a7869a7d39033d23480d7cfb18
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188126
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>