tdf#167476 cui a11y: Show border preset selection while focused
As discussed in tdf#167476, the logic for the "Presets" control
is that it doesn't have a real permanent selection state beyond
the point in time when it has focus, i.e. it is primarily meant
to be used to allow a one-time selection of borders that can then
be tweaked later on, in the "User-defined" control below.
The preset used to initialize those borders is then (more or less)
discarded and the borders shown in the "User-defined" control are used
from that point on to manage state. (Adding or removing a border in
that "User-defined" control would invalidate any preset previously
selected.)

However, while the control has focus, the currently selected/focused
entry should be clearly indicated, which is not the case when
selection is cleared (in particular visible with gtk3's implementation
of weld::IconView).
Therefore, don't clear selection right when it happens,
but when focus moves out of the control.

Change-Id: Id92e06cb07b7836c1826c966e81b8ad2ab488c04
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188138
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins