IconView/ListBox a11y: Don't support XAccessibleText w/o text
AccessibleListBoxEntry is used for both, entries that
actually can contain text as well as for ones that don't
actually support/have text (typically IconView entries).

So far, the XAccessibleText interface was always
reported as supported anyway.

With Orca on Linux with the qt6 VCL plugin, this resulted
in Orca announcing "blank" instead of the accessible name
e.g. for items in Writer's "Insert" -> "Fontwork" dialog
(when used without SAL_VCL_QT_USE_WELDED_WIDGETS=1, i.e.
using the VCL implementation, not the native Qt one) or
for items in the IconViews newly used in the
"Format" -> "Page Style" dialog, tab "Borders" since

    commit 4d6429c0368f9b2ec796dd8468defbee194a8b4c
    Date:   Thu Jul 17 14:52:51 2025 +0530

        tdf#167536 Switch to IconView for presets and shadows in border page

despite follow-up commit

    commit 7fc56525ed18de950d3892d69942f8dc8d15f22b
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Mon Jul 21 23:49:37 2025 +0200

        tdf#167536 cui: Set a11y names for IconView items in borders tab page

to set the a11y names again (which had made announcement
work again for the gtk3 VCL plugin that uses a native
GtkIconView, but not qt6 using the VCL implementation).

This is because Orca gives higher priority to text provided
using the AT-SPI Text interface (to which the XAccessibleText
interface is bridged on Linux with qt6).

Don't report support for XAccessibleText any more if the entry
doesn't contain any text item, which makes Orca announce the
accessible names again, as it used to do for the ValueSet items
used before the above-mentioned 4d6429c0368f9b2ec796dd8468defbee194a8b4c
since Orca commit [1]

    commit db2395fbca4fff4dd452ef8369cb8f253eec20e2
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Feb 25 14:06:12 2025 +0100

        soffice: Don't announce "blank" as text for non-text objects

        While "blank" may be useful as an information that a
        text object (like a LibreOffice Writer paragraph)
        does not contain any text, it's not useful for non-text
        objects. Therefore, only try to get the text content
        of the current line and return "blank" if it's empty
        in the soffice SpeechGenerator._generate_text_line method
        if the object supports the AtSpiText interface.

        Otherwise, this causes the

            result += (self._generate_text_line(obj, **args) or \
                self._generate_accessible_label_and_name(obj, **args))

        logic in SpeechGenerator._generate_list_item to use
        "blank" for non-text items instead of their accessible label/name.

        Sample scenario:

        1) start LibreOffice Writer (gtk3 or qt6 variant,
          tested with current LO development version as of
          commit 8b998ed161f0c7e72530e5f02aa9be13edb4f703)
        2) open the "Format" -> "Paragraph" dialog
        3) go to the "Borders" tab
        4) move focus to the "Shadow Style" section, list of
          items labelled by the "Position" label
        5) move between the items using the arrow keys

        Result without this commit in place:
        Orca announces "blank" for each of the items.

        Result with this commit in place:
        Orca announces the actual name
        ("None", "Cast Shadow to Bottom Right",...)..

[1] https://gitlab.gnome.org/GNOME/orca/-/commit/db2395fbca4fff4dd452ef8369cb8f253eec20e2

Change-Id: Ifd1f6034e00cea9f56883830483c9b2bbc874aec
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188162
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins