IconView/ListBox a11y: Ignore non-children for getAccessibleAtPoint
XAccessibleComponent::getAccessibleAtPoint is supposed
to return the *direct child* at the given relative
coordinates, or an empty reference, see the doc.

Therefore, ignore the accessible for any entry returned
by SvTreeListBox::GetEntry in
AccessibleListBoxEntry::getAccessibleAtPoint
that is not an actual child (i.e. that doesn't have `this`
as parent).

This fixes a freeze (due to some logic querying the same object
over and over again) seen when using Accerciser's "Inspect object
under mouse" feature while the mouse is over one of the elements
in Math's operator sidebar (e.g. over the "+ Sign") when using
the qt6 VCL plugin on Linux.

Instead, the area of the element hovered actually gets highlighted
on screen now.

Change-Id: Ifda410eb2d8220361bfe161e9dcb18a4120a6806
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188164
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>