IconView a11y: Don't notify of entry focus on mouseover
    commit 2a28ebeef5ea3e2b01d836a7233d2316b765bf38
    Date:   Wed Jun 1 11:18:26 2022 +0300

        Accessibility for IconView

had introduced AccessibleIconView and implemented
sending an ACTIVE_DESCENDANT_CHANGED event when
the mouse hovers over one of the entries, in order
to make the NVDA screen reader on Windows announce
that item.

However, ACTIVE_DESCENDANT_CHANGED events are to
indicate that the corresponding entry received
keyboard focus, which isn't the case, so sending
that event could break focus tracking in assistive
technologies.

As described in more detail in [1], the non-announcement
of the accessible name of the hovered IconView entry
was caused by NVDA trying to retrieve a text representation
via the IAccessibleText interface and not making use of the
accessible name, which is effectively the same root cause
as described for Orca's non-announcement for focused IconView
entries with qt6 on Linux and fixed in previous commit

    Change-Id: Ifd1f6034e00cea9f56883830483c9b2bbc874aec
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Jul 22 12:43:29 2025 +0200

        IconView/ListBox a11y: Don't support XAccessibleText w/o text

Together with

    Change-Id: Ifda410eb2d8220361bfe161e9dcb18a4120a6806
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Jul 22 13:58:20 2025 +0200

        IconView/ListBox a11y: Ignore non-children for getAccessibleAtPoint

, the IconView items in Math's operator sidebar are now announced
on mouseover as expected by NVDA even without the workaround to send
ACTIVE_DESCENDANT_CHANGED events (if mouse tracking is enabled in
NVDA settings).

Therefore, drop that workaround and the now unnecessary
AccessibleIconView subclass altogether.

[1] https://gerrit.libreoffice.org/c/core/+/135226/comment/bd02badb_6aa060a6/

Change-Id: If388ac0e901f3581fb8cacb6e57317068bc14f9e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188165
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>