Move variable declaration to the place it's actually used
This shows why. The string, created just to be re-initialized later, was
mistakenly used in a message constructed for an exception; and that went
unnoticed just because of that early declaration. Not a critical error,
but made the diagnostic less useful.

Was that way ever since commit 7c25b2baae4c8f45c7a588598b3af0a8aae3fc39
(INTEGRATION: CWS jl49 (1.1.2); FILE ADDED, 2006-12-20).

Change-Id: I253a99fafa6720faadce0b2e4dc0ff793c243530
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188196
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins