tdf#130857 qt dnd: Align QDragEnterEvent handling with other DnD events
For all of the other Drag'n'Drop events, QtWidget
forwards them to the QtFrame and that one implements
the logic, but for the QDragEnterEvent in
QtWidget::dragEnterEvent, QtWidget was handling it
directly.

However, then the handling of the QDragMoveEvent
in QtFrame::handleDragMove (called
from QtWidget::drageMoveEvent) was still handling
the first time it was called specially, by calling
QtDropTarget::dragEnter and setting QtFrame::m_bInDrag
to true, i.e. it was implementing logic that sounds
actually related to the QDragEnterEvent, not the
QDragMoveEvent.

Instead, also forward the QDragEnterEvent from the
QtWidget to its QtFrame and consider that one the
start of the drag action (call QtDropTarget::dragEnter
there instead) and move the previous logic from
QtWidget::dragEnterEvent to accept the event there,
too.

The QDragEnterEvent doc [1] says:

> A widget must accept this event in order to receive
> the drag move events that are sent while the drag and
> drop action is in progress. The drag enter event is
> always immediately followed by a drag move event.

Therefore, even without this change in place,
QtDropTarget::dragEnter would be called instantly
after the processing of the QDragEnterEvent when
the subsequent QDragMoveEvent was handled, but
doing that when the actual enter event occurs
seems more logical (and straightforward) to me, as
each of the Qt events now results in the corresponding
method to notify listeners of the QtDropTarget getting
called:

* QDragEnterEvent: QtDropTarget::dragEnter
* QDragMoveEvent: QtDropTarget::dragOver
* QDropEvent: QtDropTarget::drop
* QDragLeaveEvent: QtDropTarget::dragExit

This also means that QtFrame::m_bInDrag is no longer
needed to distinguis the first QDragMoveEvent
(directly following the QDragEnterEvent) from
later ones, so drop that member.

No change in behavior expected or seen in a quick
test of dragging and dropping text from Kate or Gedit
to Writer and vice versa or between two Writer windows.

[1] https://doc.qt.io/qt-6/qdragenterevent.html#details

Change-Id: I3878ca79aaf81af7fd36335bf2f90ccf0be9bf63
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188229
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>