tdf#130857 qt dnd: Let helpers multiply scale factor
Let the toVclDropTargetDragEnterEvent and
toVclDropTargetDropEvent take the scale
factor as a param and multiply it to the
x and y coords in there instead of doing so in a
separate step in QtFrame after calling the helper
functions. This will allow moving more logic from
QtFrame to QtDropTarget in an upcoming commit
to prepare for implementing
QtInstanceWidget::get_drop_target etc.
without duplicating that logic.

Change-Id: I152614be5db73378e4d7ec965975a68c19203a8b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188230
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>