tdf#130857 qt dnd: Move some logic from QtFrame to QtDropTarget
Move most of the logic to handle the various Qt drop
event types from QtFrame to QtDropTarget.
This will allow reusing the logic in an upcoming
commit implementing QtInstanceWidget::get_drop_target.

The order in QtFrame::handleDrop slightly changes
as the drag source is now notified after the QEvent
has been accepted, but that shouldn't be relevant.

No change in behavior expected or seen in a quick
test of dragging and dropping text from Kate or Gedit
to Writer and vice versa or between two Writer windows.

Change-Id: I2b4df4459028b7c2b1da300306e12f032a7605ff
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188231
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins