tdf#130857 qt weld dnd: Implement drop support
Implement the "Drop" side of Drag'n'Drop support
by implementing QtInstanceWidget::get_drop_target
and handling the corresponding events in
QtInstanceWidget::eventFilter.

The relevant logic was moved to QtDropTarget in
previous commits up to

    Change-Id: I2b4df4459028b7c2b1da300306e12f032a7605ff
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Jul 23 13:36:22 2025 +0200

        tdf#130857 qt dnd: Move some logic from QtFrame to QtDropTarget

in order to be able to reuse it from here.

In a WIP branch for supporting the "Tools" -> "SQL" dialog
in Base, this allows to insert text into the
"Command to execute" field by selecting it in Kate or Gedit
and using drag'n'drop to get it into the LO dialog.

Change-Id: I636fbcf1bc8aba67b8e32f81291efe10bc08ac9c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188232
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>