qt dnd: Move some logic from QtTools to QtDropTarget
These are only used in QtDragAndDrop.cxx and
are not very generic, so move them to the latter
source file instead.

Also change the helper functions to create a VCL
drop event from a QEvent to helper methods
in QtDropTarget and use `this` directly instead
of them being static functions that take a QtDropTarget*
param.

Most of the logic was previously moved to QtTools in

    commit cc4dbbf714faef42cd17ccd7fcfc240d3e5aa977
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Jul 1 16:49:38 2025 +0200

        tdf#130857 qt: Extract helpers to convert Qt drag/drop to LO events

with the idea to share it between QtFrame and
QtInstanceWidget, but the logic was now eventually
moved from QtFrame to QtDropTarget in

    Change-Id: I2b4df4459028b7c2b1da300306e12f032a7605ff
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Jul 23 13:36:22 2025 +0200

        tdf#130857 qt dnd: Move some logic from QtFrame to QtDropTarget

instead and that is reused by QtInstanceWidget, so it
is no longer needed elsewhere.

Change-Id: I2303d88d74bb1da4d1871396075737c1f0b6866c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188233
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>