Keep ViewShells sorted in most recently used order
SfxViewShell::GetCurrent() is the current shell. Tt would be convenient
to have the list that GetFirst/GetNext work with, in MRU order to be
consistent with that, so it can be used to get the most recently current
viewshell that matches a specific criteria.

Also keep ViewFrames in MRU order for consistency with MRU ViewShell
ordering.

Change-Id: I15126242a4895bb90d8bcf34d91e917a217f4204
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188333
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>