tdf#167552 sw: Disable DOC handling of fDntBlnSbDbWid compat flag
This change partially reverts the following commit:

commit 6818bc55ff248c59f12b2e090139eff30fe949dd
    tdf#88908 sw: Add BalanceSpacesAndIdeographicSpaces compat option

Specifically, this change disables the actual handling of the compat
flag for DOC files. It continues to work correctly for DOCX files.

We were parsing the compatibility flag correctly. However, there is some
undocumented and as-yet unknown mechanism Word uses to control whether
or not the fDntBlnSbDbWid value is effective for a particular document.
I looked at all of the likely Dop fields across the sample documents I
have available, but I couldn't see a pattern. Possibly this is based on
a version field located elsewhere, or maybe it's even based on the size
of the Dop structure itself.

After spending time trying to reverse engineer this, I'm growing
concerned that this hints at a general Copts masking mechanism built
into Word that we don't implement, rather than something that affects
this one compat flag in particular.

Change-Id: I021c636c0a9c72c5489960221f5ecec7823bf566
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188247
Reviewed-by: Jonathan Clark <jonathan@libreoffice.org>
Tested-by: Jenkins